Code refactoring, removing a lot of duplicate code sections
