+++ src/main/java/jenkins/plugins/linkedjobs/actions/LabelLinkedJobsAction.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License
  * 
- * Copyright (C) 2014 Dominique Brice
+ * Copyright (C) 2014, 2015 Dominique Brice
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -29,14 +29,7 @@ import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
 
-import jenkins.plugins.linkedjobs.helpers.TriggeredJobsHelper;
 import jenkins.plugins.linkedjobs.model.JobsGroup;
-import jenkins.plugins.linkedjobs.model.TriggeredJob;
-import jenkins.plugins.linkedjobs.settings.GlobalSettings;
-import jenkins.model.Jenkins;
-import hudson.model.Action;
-import hudson.model.TopLevelItem;
-import hudson.model.AbstractProject;
 import hudson.model.Label;
 import hudson.model.labels.LabelAtom;
 
@@ -48,7 +41,7 @@ import hudson.model.labels.LabelAtom;
  * @author dominiquebrice
  *
  */
-public class LabelLinkedJobsAction implements Action {
+public class LabelLinkedJobsAction extends AbstractLinkedJobsAction {
     
     /**
      * The label associated to this action
@@ -61,32 +54,13 @@ public class LabelLinkedJobsAction implements Action {
         this.label = labelAtom;
     }
 
-    public String getIconFileName() {
-        return "search.png";
-    }
-
-    // name of the additional link in the left-side menu for labels
-    public String getDisplayName() {
-        return "Linked Jobs";
-    }
-
     public String getTitle() {
         return label.getDisplayName() + " " + getDisplayName();
-    }
-
-    // relative URL used for the additional link in the left-side menu for
-    // labels
-    public String getUrlName() {
-        return "linkedjobs";
     }
 
     // will be called by my friend jelly via ${it.label}. Of course!
     public LabelAtom getLabel() {
         return this.label;
-    }
-
-    public boolean getDetailedView() {
-        return GlobalSettings.get().getDetailedView();
     }
 
     /**
@@ -97,58 +71,10 @@ public class LabelLinkedJobsAction implements Action {
      * @return a list of jobs grouped by assigned label
      */
     public List<JobsGroup> getJobsGroups() {
-        HashMap<Label, JobsGroup> tmpResult = new HashMap<Label, JobsGroup>();
-
-        // this loop is directly inspired from hudson.model.Label.getTiedJobs()
-        for (AbstractProject<?, ?> job : Jenkins.getInstance().getAllItems(
-                AbstractProject.class)) {
-            if (job instanceof TopLevelItem) {
-                Label jobLabel = job.getAssignedLabel();
-                // condition for a job to be listed is simply to use the label,
-                // regardless of the actual meaning of the expression
-                // In other words, jobs with assigned label "windows&&!jdk7"
-                // will show up on the jdk7 "linked jobs" page
-                if (isAssignedLabelLinked(jobLabel)) {
-                    JobsGroup matchingJobGroup = tmpResult.get(jobLabel);
-                    if (matchingJobGroup == null) {
-                        matchingJobGroup = new JobsGroup(jobLabel);
-                        tmpResult.put(jobLabel, matchingJobGroup);
-                    }
-                    matchingJobGroup.addJob(job);
-                }
-            }
-        }
-        
-        // then browse list of triggered jobs
-        HashMap<Label, HashMap<AbstractProject<?,?>, TriggeredJob>> triggeredJobsByLabel =
-                new HashMap<Label, HashMap<AbstractProject<?,?>, TriggeredJob>>();
-        TriggeredJobsHelper.populateTriggeredJobs(triggeredJobsByLabel);
-        for (Label label : triggeredJobsByLabel.keySet()) {
-            if (isAssignedLabelLinked(label)) {
-                JobsGroup matchingJobGroup = tmpResult.get(label);
-                if (matchingJobGroup == null) {
-                    matchingJobGroup = new JobsGroup(label);
-                    tmpResult.put(label, matchingJobGroup);
-                }
-                // get the list of all triggered jobs
-                matchingJobGroup.addTriggeredJobs(triggeredJobsByLabel.get(label).values());
-            }
-        }
-        
-        // then browse list of jobs with a Label parameter with a default value
-        HashMap<Label, List<AbstractProject<?,?>>> jobsByDefaultLabel = new HashMap<Label, List<AbstractProject<?,?>>>();
-        TriggeredJobsHelper.populateJobsWithLabelDefaultValue(jobsByDefaultLabel);
-        for (Label label : jobsByDefaultLabel.keySet()) {
-            if (isAssignedLabelLinked(label)) {
-                JobsGroup matchJobsGroup = tmpResult.get(label);
-                if (matchJobsGroup == null) {
-                    matchJobsGroup = new JobsGroup(label);
-                    tmpResult.put(label, matchJobsGroup);
-                }
-                matchJobsGroup.addJobsWithDefaultValue(jobsByDefaultLabel.get(label));
-            }
-        }
-
+        return buildJobsGroups();
+    }
+    
+    protected List<JobsGroup> buildResult(HashMap<Label, JobsGroup> tmpResult) {
         ArrayList<JobsGroup> result = new ArrayList<JobsGroup>(tmpResult.size());
         // keep track from where we must finally sort the result
         int fromIndex = 0;
@@ -161,11 +87,14 @@ public class LabelLinkedJobsAction implements Action {
         result.addAll(tmpResult.values());
         // and order them
         Collections.sort(result.subList(fromIndex, result.size()));
-
         return result;
     }
     
-    protected boolean isAssignedLabelLinked(Label jobLabel) {
+    protected boolean isLabelRelevant(Label jobLabel) {
+        // condition for a job to be listed is simply to use the label,
+        // regardless of the actual meaning of the expression
+        // In other words, jobs with assigned label "windows&&!jdk7"
+        // will show up on the jdk7 "linked jobs" page
         return jobLabel != null && jobLabel.listAtoms().contains(this.label);
     }
 }
