+++ src/main/java/jenkins/plugins/linkedjobs/model/LabelAtomData.java
@@ -1,7 +1,7 @@
 /*
  * The MIT License
  * 
- * Copyright (C) 2014 Dominique Brice
+ * Copyright (C) 2014, 2015 Dominique Brice
  * 
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -25,53 +25,27 @@
 package jenkins.plugins.linkedjobs.model;
 
 import java.util.ArrayList;
-import java.util.Collection;
 
 import jenkins.model.Jenkins;
 import jenkins.plugins.linkedjobs.actions.LabelLinkedJobsAction;
-import hudson.model.AbstractProject;
 import hudson.model.Node;
 import hudson.model.labels.LabelAtom;
 import hudson.util.VersionNumber;
 
-public class LabelAtomData implements Comparable<LabelAtomData> {
+public class LabelAtomData extends AbstractJobsGroup implements Comparable<LabelAtomData> {
 
     private final LabelAtom labelAtom;
-    // list of jobs sharing this label
-    private ArrayList<AbstractProject<?, ?>> jobs;
     // list all nodes defining this label
     private ArrayList<Node> nodes;
-    // list of triggered jobs sharing this label
-    private ArrayList<TriggeredJob> triggeredJobs;
-    // list of jobs using this label as a default value for their Label parameter
-    private ArrayList<AbstractProject<?, ?>> jobsWithLabelDefaultValue;
 
     public LabelAtomData(LabelAtom l) {
+        super();
         labelAtom = l;
-        jobs = new ArrayList<AbstractProject<?, ?>>();
-        triggeredJobs = new ArrayList<TriggeredJob>();
-        jobsWithLabelDefaultValue = new ArrayList<AbstractProject<?,?>>();
         nodes = new ArrayList<Node>();
-    }
-
-    public void add(AbstractProject<?, ?> job) {
-        jobs.add(job);
-    }
-    
-    public void addTriggeredJobs(Collection<TriggeredJob> jobs) {
-        triggeredJobs.addAll(jobs);
-    }
-    
-    public void addJobsWithDefaultValue(Collection<AbstractProject<?, ?>> jobs) {
-        jobsWithLabelDefaultValue.addAll(jobs);
     }
 
     public void add(Node n) {
         nodes.add(n);
-    }
-    
-    public LabelAtom getLabelAtom() {
-        return labelAtom;
     }
     
     /************************************
@@ -94,20 +68,6 @@ public class LabelAtomData implements Comparable<LabelAtomData> {
     
     public String getLabelURL() {
         return labelAtom.getUrl();
-    }
-    
-    public int getJobsCount() {
-        return jobs.size();
-    }
-
-    public int getTriggeredJobsCount() {
-        return triggeredJobs.size();
-    }
-    
-    // return the number of jobs that uses this LabelAtom
-    // as (part of) their default value for a Label parameter
-    public int getJobsWithLabelDefaultValueCount() {
-        return jobsWithLabelDefaultValue.size();
     }
     
     public int getNodesCount() {
